4Z33DKP drop the dead React imports the split copied, and give CommitDiffState one home
The split left `import React` in five new files where nothing referenced the
namespace, and had CommitRow reach back into IssueCommits for a type that
lib/use-issue-detail already declared identically — a cycle between the two
components. Both now take the lib copy. The scrubber barrel also stops
re-exporting five names that only its own submodules use.
